Send Phone Msg with Twilio
The code sent a Msg using the Twilio library,  it uses and sends to my cellphone the information for 3 APIS: the location of the ISS, one Harry Potter Character, and the actual Weather of my city. 

To run it without error, use your own twilio key and introduce your phone.   

Don't forget to fork it!

https://replit.com/join/cofgktgejo-gabrielaleon3